requirements.txt
adal==1.2.7
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.6
ansiwrap==0.8.4
anyio==3.5.0
apache-airflow==2.2.4
apache-airflow-providers-ftp==2.1.0
apache-airflow-providers-http==2.1.0
apache-airflow-providers-imap==2.2.1
apache-airflow-providers-papermill==2.2.2
apache-airflow-providers-sqlite==2.1.1
apispec==3.3.2
argcomplete==2.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==20.3.0
azure-core==1.23.0
azure-datalake-store==0.0.52
azure-storage-blob==12.10.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
blinker==1.4
boto3==1.21.23
botocore==1.24.23
cachelib==0.6.0
cachetools==5.0.0
cattrs==1.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
clickclick==20.10.2
cloudpickle==2.0.0
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
connexion==2.12.0
croniter==1.3.4
cryptography==36.0.1
cycler==0.11.0
databricks-cli==0.16.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
dnspython==2.2.1
docker==5.0.3
docutils==0.16
email-validator==1.1.3
entrypoints==0.4
executing==0.8.3
Flask==1.1.4
Flask-AppBuilder==3.4.4
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fonttools==4.31.2
frozenlist==1.3.0
fsspec==2022.2.0
gcsfs==2022.2.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.7.1
google-auth==2.6.2
google-auth-oauthlib==0.5.1
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
graphviz==0.19.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
inflection==0.5.1
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
iso8601==1.0.2
isodate==0.6.1
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
jmespath==1.0.0
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyter-server==1.15.3
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.0
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mlflow==1.24.0
msrest==0.6.21
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.3.6
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.9
notebook-shim==0.1.0
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
papermill==2.3.4
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prison==0.2.1
prometheus-client==0.13.1
prometheus-flask-exporter==0.19.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
querystring-parser==1.2.4
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.0.0
rsa==4.8
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.8.0
scrapbook==0.5.0
seaborn==0.11.2
Send2Trash==1.8.0
setproctitle==1.2.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
stack-data==0.2.0
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
textwrap3==0.9.2
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typing-extensions==4.1.1
unicodecsv==0.14.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==1.0.1
widgetsnbextension==3.6.0
wrapt==1.14.0
WTForms==2.3.3
yarl==1.7.2
zipp==3.7.0
